backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
groq>=0.9.0
qdrant-client>=1.9.0
sentence-transformers>=2.7.0
pymupdf>=1.23.0
python-multipart>=0.0.9
